refactor: remove old NostrEvents component
